+++ wakamiti-api/src/test/java/es/iti/wakamiti/api/matcher/CharSequenceLengthMatcherTest.java
@@ -3,12 +3,14 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at https://mozilla.org/MPL/2.0/.
  */
-package es.iti.wakamiti.rest.matcher;
+package es.iti.wakamiti.api.matcher;
+
 
 import org.junit.Test;
 
 import static org.hamcrest.MatcherAssert.assertThat;
 import static org.hamcrest.Matchers.equalTo;
+
 
 public class CharSequenceLengthMatcherTest {